+++ src/glyphs/row.cpp
@@ -44,7 +44,7 @@ void Row::InsertBack(std::list<GlyphPtr>&& list) {
     return Insert(insertPosition, std::move(list));
 }
 
-void Row::Insert(size_t insertPosition, std::list<GlyphPtr>&& itemsToInsert) {
+void Row::Insert(int insertPosition, std::list<GlyphPtr>&& itemsToInsert) {
     auto updateElements = [&](int offset){
       for (auto& it: itemsToInsert) {
           it->SetPosition(offset, y);
